refactor(structure): flattened folder structure & moved custom validation service to shared services
